+++ src/Components/NavBar/NavBar.tsx
@@ -22,7 +22,6 @@ import {
   MobileNavMenu,
   MobileToggleIcon,
   MoreNavMenu,
-  NewMobileNavMenu,
   NotificationsMenu,
   UserMenu,
 } from "./Menus"
@@ -54,9 +53,7 @@ export const NavBar: React.FC = track(
   const { xs, sm } = useMedia()
   const isMobile = xs || sm
   const isLoggedIn = Boolean(user)
-  const canViewNewMobileNav = Boolean(
-    user?.lab_features?.includes("Updated Navigation")
-  )
+
   const getNotificationCount = () => cookie.get("notification-count") || 0
 
   // Close mobile menu if dragging window from small size to desktop
@@ -234,11 +231,7 @@ export const NavBar: React.FC = track(
       {showMobileMenu && (
         <>
           <MobileNavCover onClick={() => toggleMobileNav(false)} />
-          {canViewNewMobileNav ? (
-            <NewMobileNavMenu isOpen={showMobileMenu} menuData={menuData} />
-          ) : (
-            <MobileNavMenu onNavItemClick={() => toggleMobileNav(false)} />
-          )}
+          <MobileNavMenu isOpen={showMobileMenu} menuData={menuData} />
         </>
       )}
     </header>
